Float power fix

+++ src/loss.cpp
@@ -13,7 +13,7 @@ void MeanSquareError::compute_derivative(Matrix &y_pred, Matrix &y_true, Matrix 
 }
 
 float MeanSquareError::compute_point(float y_pred, float y_true) const {
-    return pow(y_pred - y_true, 2.0f);
+    return powf(y_pred - y_true, 2.0f);
 }
 
 float MeanSquareError::compute_derivative_point(float y_pred, float y_true) const {
